MEDICAL DIAGNOSIS WITH DEEP LEARNING/requirements.txt
tensorflow==2.16.1
numpy<2.0
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
pillow==10.3.0
opencv-python-headless==4.9.0.80
streamlit==1.36.0
plotly==5.22.0
jupyterlab==4.2.1
ipywidgets==8.1.2
